feat(inquiries): full back-and-front integration for inquiries and responsive UI
This commit includes:
- Merge latest backend version
- Create InquiryForm component with title, description, and property selector
- Add client-side validation and error alerts
- Create AppointmentComponent (UI only)
- Build InquiryPanel for streamlined inquiry navigation
- Update form formatting and file structure for consistency
- Create ManagerPanel with redirects to specific panels for better component management
- Fetch and render user’s turns in User Profile
- Display inquiry status badges
- Add data table showing all inquiries with filters: All / General / Property / Status

Refs: #33, #35, #36

+++ backend/ms-properties/src/main/java/pi/ms_properties/dto/PropertyDTO.java
@@ -7,6 +7,7 @@ import pi.ms_properties.domain.Amenity;
 import pi.ms_properties.domain.Image;
 import pi.ms_properties.domain.Type;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.Set;
 
@@ -23,9 +24,9 @@ public class PropertyDTO {
     private Float bedrooms;
     private Float area;
     private Float coveredArea;
-    private Float price;
+    private BigDecimal price;
     private Boolean showPrice;
-    private Float expenses;
+    private BigDecimal expenses;
     private Boolean credit;
     private Boolean financing;
     private String description;
